implemented rotation
I've generated only the images with rotation of 45° and 90°.  I decided to use the expand = True , this change the size of the image for allow to see the entire image. If we want we can set it to False to keep fixed the of #pixels, however this cut the image's corners.